MCStatus/PythonProxy/requirements.txt
mcstatus==2.1
simpleyaml==1.0

